Change configuration validation controller
